requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
huggingface-hub==0.36.0
langchain-huggingface==1.0.0
langchain==1.0.3
langchain-core==1.0.2
langchain-text-splitters==1.0.0
pinecone==6.0.1
python-dotenv==1.0.0
groq==0.33.0
sentence-transformers==5.1.2
beautifulsoup4==4.14.2
lxml==6.0.2
schedule==1.2.2
requests==2.32.5
